Addition of comma and removing the extra braces
The function dump_picture_info() has extra braces in the profiles
HEVC_MAIN_444 and VP9. Because of this, the additoinal braces are
added in the trace file generated by the applications like ffmpeg,
mplayer and so on. The client of this generated trace file, like
the python scripts that read this trace file, run into into the
problem of wrongly reading the picture info or incomplete reading
of picture info.

Also, there was a comma missing in the VP9 profile in the
dump_picture_info(). This again creates the problem of generating
a trace file without a comma, which again leads to wrongly reading
the picture info.

To avoid this problem of wrongly reading or incomplete reading of
picture info, extra braces have been removed and an additional
comma is introduced in the dump_picture_info().